requirements.txt
pytest~=6.2.4
requests~=2.26.0
python-decouple~=3.5
setuptools~=58.0.4